Make pretty results. Added asserts to the crypto API return values. Modified posix radio.cpp to use key material. Modified radio_platform.cpp to use key material. Added crypto_platform.cpp and hmac_sha256 to radio cmake file. Minor bug fix to key manager.

+++ include/openthread/platform/crypto.h
@@ -137,7 +137,7 @@ otError otPlatCryptoImportKey(psa_key_id_t *        aKeyId,
  * @retval OT_ERROR_INVALID_ARGS  @p aBuffer was NULL
  *
  */
-otError otPlatCryptoExportKey(psa_key_id_t aKeyId, uint8_t *aBuffer, uint8_t aBufferLen, size_t *aKeyLen);
+otError otPlatCryptoExportKey(psa_key_id_t aKeyId, uint8_t *aBuffer, uint16_t aBufferLen, size_t *aKeyLen);
 
 /**
  * Destroy a key stored in PSA ITS.
